requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pydantic==2.5.0
requests
firecrawl-py==2.16.5
